pyproject.toml
[project]
name = "polyai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.13"
dependencies = [
    "black>=26.3.1",
    "grpcio>=1.78.0",
    "gymnasium==1.2.2",
    "ray[rllib]>=2.54.0",
    "websockets>=16.0",
]

[dependency-groups]
dev = [
    "maturin>=1.12.6",
]
